trigger VM cache invalidation after tx complete
