+++ playwright-tests/tests/dontaskagain.spec.js
@@ -3,10 +3,14 @@ import { pauseIfVideoRecording } from "../testUtils.js";
 import {
   getDontAskAgainCacheValues,
   setDontAskAgainCacheValues,
+  findKeysInCache
 } from "../util/cache.js";
 import {
   modifySocialNearGetRPCResponsesInsteadOfGettingWidgetsFromBOSLoader
 } from "../util/bos-loader.js";
+import { mockTransactionSubmitRPCResponses } from '../util/transaction.js';
+
+const RECEIVER_ID = 'devgovgigs.near';
 
 test.describe("Wallet is connected with devhub access key", () => {
   test.use({
@@ -47,97 +51,22 @@ test.describe("Wallet is connected with devhub access key", () => {
     await submitbutton.scrollIntoViewIfNeeded();
     await pauseIfVideoRecording(page);
 
-    await page.route("https://rpc.mainnet.near.org/", async (route) => {
-      const request = await route.request();
-
-      const requestPostData = request.postDataJSON();
-      if (
-        requestPostData.params &&
-        requestPostData.params.request_type === "view_access_key_list"
-      ) {
-
-        const response = await route.fetch();
-        const json = await response.json();
-        json.result.keys = [{
-          "access_key": {
-            "nonce": 109629226000005, "permission": {
-              "FunctionCall": {
-                "allowance": "241917078840755500000000", "method_names": [],
-                "receiver_id": "devgovgigs.near"
-              }
-            }
-          },
-          "public_key": "ed25519:EQr7NpVYFu1XcVZ23Lb4Ga3KbDQgrYeTMTgBsYa26Bne"
-        }];
-
-        console.log("Replacing RPC response when listing access keys", JSON.stringify(json));
-        await route.fulfill({ response, json });
-      } else if (requestPostData.params &&
-        requestPostData.params.request_type === "view_access_key") {
-        const response = await route.fetch();
-        const json = await response.json();
-
-        json.result = {
-          "nonce": 85,
-          "permission": {
-            "FunctionCall": {
-              "allowance": "241917078840755500000000",
-              "receiver_id": "devgovgigs.near",
-              "method_names": []
-            }
-          },
-          "block_height": 19884918,
-          "block_hash": "GGJQ8yjmo7aEoj8ZpAhGehnq9BSWFx4xswHYzDwwAP2n"
-        };
-
-        console.log("Replacing RPC response when viewing access key", JSON.stringify(json));
-        await route.fulfill({ response, json });
-      } else if (requestPostData.method == 'broadcast_tx_commit') {
-        console.log("Replacing RPC response when broadcasting tx");
-        await route.fulfill({
-          json: {
-            "jsonrpc": "2.0",
-            "result": {
-              "status": {
-                "SuccessValue": ""
-              },
-              "transaction_outcome": {
-                "proof": [],
-                "block_hash": "9MzuZrRPW1BGpFnZJUJg6SzCrixPpJDfjsNeUobRXsLe",
-                "id": "ASS7oYwGiem9HaNwJe6vS2kznx2CxueKDvU9BAYJRjNR",
-                "outcome": {
-                  "logs": [],
-                  "receipt_ids": ["BLV2q6p8DX7pVgXRtGtBkyUNrnqkNyU7iSksXG7BjVZh"],
-                  "gas_burnt": 223182562500,
-                  "tokens_burnt": "22318256250000000000",
-                  "executor_id": "sender.testnet",
-                  "status": {
-                    "SuccessReceiptId": "BLV2q6p8DX7pVgXRtGtBkyUNrnqkNyU7iSksXG7BjVZh"
-                  }
-                }
-              },
-              "receipts_outcome": [
-
-              ]
-            }
-          }
-        });
-      } else {
-        console.log('unmodified', JSON.stringify(requestPostData));
-        await route.continue();
-      }
-    });
+    const cachedValues = await findKeysInCache(page, RECEIVER_ID);
+    console.log('cached values', cachedValues);
+    await mockTransactionSubmitRPCResponses(page, RECEIVER_ID);
 
     await submitbutton.click();
     await pauseIfVideoRecording(page);
     await submitbutton.waitFor({ state: "detached", timeout: 500 });
-    const callContractToast = await page.getByText("Calling contract devgovgigs.near with method add_post");
+
+    const callContractToast = await page.getByText(`Calling contract ${RECEIVER_ID} with method add_post`);
     expect(
       callContractToast
         .isVisible()
     ).toBeTruthy();    
     await callContractToast.waitFor({state: "detached"});
     await pauseIfVideoRecording(page);
+    
     await page.waitForTimeout(5000);
   });
